requirements.txt
livekit-agents>=1.3.10,<2.0.0
av>=14.0.0
livekit-plugins-deepgram>=1.3.10,<2.0.0
livekit-plugins-cartesia>=1.3.10,<2.0.0
livekit-plugins-elevenlabs>=1.3.10,<2.0.0
livekit-plugins-openai>=1.3.10,<2.0.0
livekit-plugins-silero>=1.3.10,<2.0.0
livekit-plugins-turn-detector>=1.3.10,<2.0.0
livekit-plugins-noise-cancellation>=0.2.5
python-dotenv>=1.0.1
requests>=2.28.0
# Optional: pip install 'livekit-agents[mcp]' for MCP server integration
# For running mcp_server.py: pip install mcp


